+++ firmware/arduino/src/Sensors/MainPressureSensor/MainPressureSensor.cpp
@@ -7,9 +7,9 @@ MainPressureSensor::MainPressureSensor() {
 
 }
 
-unsigned int MainPressureSensor::read() {
+float MainPressureSensor::read() {
   // mm H2O
-  unsigned int pressure = (5.0* ((analogRead(PRESSURE_PIN)/1024.0) - 0.1) / 0.8) * 703;
+  float pressure = (5.0* ((analogRead(PRESSURE_PIN)/1024.0) - 0.1) / 0.8) * 703;
   return pressure;
 }
